+++ ecommerce/next.config.js
@@ -1,3 +1,6 @@
 module.exports = {
   reactStrictMode: true,
-}
+  images: {
+    domains: ["fakestoreapi.com"],
+  },
+};
