+++ ecommerce/pages/contacts.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function contacts() {
+    return (
+        <div>
+            <h1>Hello from Contacts</h1>
+        </div>
+    )
+}
+
+export default contacts
