+++ ecommerce/pages/orders.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function orders() {
+    return (
+        <div>
+            <h1>Hello from orders</h1>
+        </div>
+    )
+}
+
+export default orders
