+++ ecommerce/styles/Navbar.module.css
@@ -34,7 +34,6 @@
   padding-left: 28%;
   box-shadow: 0 0 0 1px #ccc;
   position: absolute;
-  bottom: 0;
   width: 100%;
   height: 60px; /* Height of the footer */
 }
