+++ ecommerce/styles/Products.module.css
@@ -2,5 +2,5 @@
   width: 100%;
   display: grid;
   grid-template-columns: 1fr 1fr 1fr;
-  gap: 1rem;
+  gap: 1rem;  
 }
